+++ src/lib/util/util.ml
@@ -66,13 +66,11 @@ type inst_kind = Normal | Forward | Backward
 
 type sat_solver =
   | Tableaux
-  | Tableaux_CDCL
   | CDCL
   | CDCL_Tableaux
 
 let pp_sat_solver ppf = function
   | Tableaux -> Format.fprintf ppf "Tableaux"
-  | Tableaux_CDCL -> Format.fprintf ppf "Tableaux-CDCL"
   | CDCL -> Format.fprintf ppf "CDCL"
   | CDCL_Tableaux -> Format.fprintf ppf "CDCL-Tableaux"